fix to support modeladmin and/or embedding in page
When the form is used outside the normal page context (ie. in modeladmin or embedded in another page) the new email recipient can't rely on the CMS page session variable to identify the parent. So it dies. It's safer to skip relationships until the parent is properly linked. This happens after the first write, which seems to happen immediately, so those fields get filled anyway.